+++ R/mnhn.R
@@ -60,39 +60,40 @@ get_mnhn <- function(
   x_buff <- sf::st_buffer(x, buffer)
 
   layer <- switch(key,
-                  "apb" = "patrinat_apb:apb",
-                  "apg" = "patrinat_apg:apg",
-                  "aphn" = "patrinat_aphn:aire_protection_habitats_naturels",
-                  "aplg" = "patrinat_aplg:apg",
-                  "bios" = "patrinat_bios:bios",
-                  "unesco" = "patrinat_bpm:Bien_patrimoine_mondial_UNESCO",
-                  "cdl" = "patrinat_cdl:conservatoire_littoral",
-                  "cen" = "patrinat_cen:cen",
-                  "geoparc" = "patrinat_geoparc:geoparc",
-                  "gsf" = "patrinat_gsf:gsf",
-                  "inpg" = "patrinat_inpg:inpg",
-                  "ospar" = "patrinat_ospar:pgsf",
-                  "pgsf" = "patrinat_pgsf:pgsf",
-                  "pn" = "patrinat_pn:parc_national",
-                  "pn2" = "patrinat_pn2:pn",
-                  "pnm" = "patrinat_pnm:pnm",
-                  "pnr" = "patrinat_pnr:pnr",
-                  "pprnn" = "patrinat_pprnn:pprnn",
-                  "ramsar" = "patrinat_ramsar:pnm",
-                  "rb" = "patrinat_rb:reserve_biologique",
-                  "ripn" = "patrinat_ripn:ripn",
-                  "rnc" = "patrinat_rnc:pnm",
-                  "rncfs" = "patrinat_rncfs:rncfs",
-                  "rnn" = "patrinat_rnn:rnn",
-                  "rnr" = "patrinat_rnr:rnr",
-                  "sc" = "patrinat_sc:sc",
-                  "sic" = "patrinat_sic:sic",
-                  "znieff1" = "patrinat_znieff1:znieff1",
-                  "znieffmer1" = "patrinat_znieff1_mer:znieff1_mer",
-                  "znieff2" = "patrinat_znieff2:znieff2",
-                  "znieffmer2" = "patrinat_znieff2_mer:znieff2_mer",
-                  "zpr" = "patrinat_zpr:zpr",
-                  "zps" = "patrinat_zps:zps"
+    "aphn" = "patrinat_aphn:aphn",
+    "apb" = "patrinat_apb:apb",
+    "apg" = "patrinat_apg:apg",
+    "aplg" = "patrinat_aplg:aplg",
+    "unesco" = "patrinat_bpm:Bien_patrimoine_mondial_UNESCO",
+    "cdl" = "patrinat_cdl:conservatoire_littoral",
+    "geoparc" = "patrinat_geoparc:geoparc",
+    "gsf" = "patrinat_gsf:gsf",
+    "inpg" = "patrinat_inpg:inpg",
+    "ospar" = "patrinat_ospar:pgsf",
+    "pgsf" = "patrinat_pgsf:pgsf",
+    "pn" = "patrinat_pn:parc_national",
+    "pn2" = "patrinat_pn2:pn",
+    "pnm" = "patrinat_pnm:pnm",
+    "pprnn" = "patrinat_pprnn:pprnn",
+    "pnr" = "patrinat_pnr:pnr",
+    "pprnr" = "patrinat_pprnr:pprnr",
+    "ripn" = "patrinat_ripn:ripn",
+    "rnc" = "patrinat_rnc:pnm",
+    "rncfs" = "patrinat_rncfs:rncfs",
+    "rnn" = "patrinat_rnn:rnn",
+    "rb" = "patrinat_rb:rb",
+    "bios" = "patrinat_bios:bios",
+    "rnr" = "patrinat_rnr:rnr",
+    "sc" = "patrinat_sc:sc",
+    "sic" = "patrinat_sic:sic",
+    "ramsar" = "patrinat_ramsar:ramsar",
+    "cen" = "patrinat_cen:cen",
+    "znieff1" = "patrinat_znieff1:znieff1",
+    "znieffmer1" = "patrinat_znieff1_mer:znieff1_mer",
+    "znieff2" = "patrinat_znieff2:znieff2",
+    "znieffmer2" = "patrinat_znieff2_mer:znieff2_mer",
+    "zpr" = "patrinat_zpr:zpr",
+    "zps" = "patrinat_zps:zps"
   )
 
   f <- happign::get_wfs(
@@ -166,21 +167,24 @@ seq_mnhn <- function(
 
     if (verbose) {cli::cli_progress_update(id = pb)}
 
-    # f mean feature in this context
-    f <- get_mnhn(parca, k, buffer = buffer)
-    if (nrow(f) > 0) {
-      f[[identifier]] <- id
-      valid <- c(valid, k)
-      seq_key <- sprintf("v.mnhn.%s.poly", k)
-      f_path <- seq_write(f, seq_key, dirname, id, verbose = FALSE, overwrite = overwrite)
-      path <- c(path, f_path)
-    } else {
-      empty <- c(empty, k)
-    }
+    try({
+      # f mean feature in this context
+      f <- get_mnhn(parca, k, buffer = buffer)
+      if (nrow(f) > 0) {
+        f[[identifier]] <- id
+        valid <- c(valid, k)
+        seq_key <- sprintf("v.mnhn.%s.poly", k)
+        f_path <- seq_write(f, seq_key, dirname, id, verbose = FALSE, overwrite = overwrite)
+        path <- c(path, f_path)
+      } else {
+        empty <- c(empty, k)
+      }
+    })
+
   }
   cli::cli_progress_done(id = pb)
 
-  if (verbose){
+  if (verbose){sequoia2
     if (length(valid) > 0) {
       cli::cli_alert_success(
         "{length(valid)} non-empty layer{?s} found: {.val {valid}}"